Move post processor test to their own file.
